web/quizbot/requirements.txt
discord.py==1.5.1
python-dotenv
